+++ internal/commands/mt.go
@@ -470,7 +470,6 @@ func init() {
 }
 
 func obtainMTCmd(context *cli.Context) error {
-
 	mt, err := obtainMT(context)
 	if err != nil {
 		return err
@@ -479,7 +478,7 @@ func obtainMTCmd(context *cli.Context) error {
 }
 
 func obtainMT(context *cli.Context) (string, error) {
-	mytoken := config.Get().Mytoken
+	mytoken := config.Get().Mytoken()
 	if mtCommand.TransferCode != "" {
 		return mytoken.Mytoken.FromTransferCode(mtCommand.TransferCode)
 	}
@@ -505,8 +504,9 @@ func obtainMT(context *cli.Context) (string, error) {
 		req.GrantType = api.GrantTypeMytoken
 		mtRes, err := mytoken.Mytoken.APIFromRequest(
 			api.MytokenFromMytokenRequest{
-				GeneralMytokenRequest: *req,
-				Mytoken:               mtGrant,
+				GeneralMytokenRequest:        *req,
+				Mytoken:                      mtGrant,
+				FailOnRestrictionsNotTighter: true,
 			},
 		)
 		if err != nil {